[TFN-333] rename db file
