+++ src/pages/api/matching.ts
@@ -1,7 +1,7 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 import { redirectTo, redirectToError, getUuidFromCookie } from './apiUtils';
 import { BasicService } from '../matching';
-import { Stop } from '../../data/dynamodb';
+import { Stop } from '../../data/auroradb';
 import { putStringInS3, UserFareStages } from '../../data/s3';
 import { isCookiesUUIDMatch } from './service/validator';
 import { MATCHING_DATA_BUCKET_NAME } from '../../constants';
