+++ tests/pages/direction.test.tsx
@@ -2,10 +2,10 @@ import * as React from 'react';
 import { shallow } from 'enzyme';
 
 import Direction, { getServerSideProps } from '../../src/pages/direction';
-import { getServiceByNocCodeAndLineName, batchGetStopsByAtcoCode } from '../../src/data/dynamodb';
+import { getServiceByNocCodeAndLineName, batchGetStopsByAtcoCode } from '../../src/data/auroradb';
 import { mockRawService, mockService, mockRawServiceWithDuplicates, getMockContext } from '../testData/mockData';
 
-jest.mock('../../src/data/dynamodb');
+jest.mock('../../src/data/auroradb');
 
 describe('pages', () => {
     describe('direction', () => {
